feat(animator): create animator
Animator takes the inputs and builds the timelines for the atoms.
Some properties are not yet configurable in the input format..
The `cz`-operation is not yet limited to nearby atoms,
and this distance is not yet configurable.
Some options from the input file are not yet implemented in the renderer