feat(animator): invariable groups and movement
Implemented the invariable group timing
Also changed the movement to allow invariable timing:
The machine-spec now sets a maximum speed.
For variable groups, the `max_velocity` is used instead of `jerk`.
For invariable groups, the largest distance determines the timing.
The other moves will calculate a jerk-value to move in that same time.