feat(parser,animator): allow targeting sets of ids
Can target multiple atoms or multiple zones in a single command.
Useful for commands like `cz` when not wanting to target a whole zone,
but a single atom is not enough due to needing an interaction target.
Also changed the `na`-converter to convert multiple targets
to a single instruction instead of multiple instructions in a group.