chore: add linking exception to crates using LGPL
LGPL requires dynamically linking the library
or providing application code to statically link with another version.
Rust usually links statically, so the current license would require
sharing the source code of all derived applications.
This exception allows usage of the concerned crates
in a fully closed source application,
as was the original intention by licensing them as LGPL.
All authors are ok with this change.